feat: made code refactor
