+++ src/app/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { AdminsModule } from 'src/admins/admins.module';
-import { AuthModule } from 'src/auth/auth.module';
-import { ClientsModule } from 'src/clients/clients.module';
 import { configuration } from 'src/config';
 import { MongoDBModule } from 'src/database';
+import { AdminsModule } from 'src/modules/admins/admins.module';
+import { AuthModule } from 'src/modules/auth/auth.module';
+import { ClientsModule } from 'src/modules/clients/clients.module';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
