+++ src/modules/clients/schemas/client-images.schema.ts
